Holy balls its now NEXT
So after 2+ weeks of debugging, I now have made this into a next site.

It's still buggy, but functional. So I'm pushing it with another update to come later this week.

Issues to fix;
Chat history is broken, leaving all chats into the same page

The miniPlayer flashes when typing, I need to remove excessive resets

some error pops up on the settings