Shared satisfied
I now have a shared page with a functioning database and people can play a shared game with a link. I will need to add the database env before the base is working, however this is a small matter. Next is making the bugsnap feature and a toolbar for the future tools I have coming.